Able to remove some faux-explicitly installed packages
After this commit c06691f2eca23e3cfa535e809f3aac554f1461c3 I was able to
run this command: pacman -D --asdeps $(pacman -Qeqg base) to set all of
the packages that appear to be explicitly installed to as dependencies.
This should be the behavior on a new arch install with how that new base
package works.