Welp, this sucks:
Python2 is disappearing and gimp has decided to remove support for plugins that need it in the latest version. Either we wait for version 3.0 of Gimp or we use other plugins, and even a flatpak image of the program (with everything included).

I abandon this package, whoever wants to continue with its maintenance is free to adopt it.

https://aur.archlinux.org/packages/python2-gimp/#comment-783877